Verbose flags in loop methods, indiv. asset plots
Added verbose keyword argument to all loop and plotting-related methods
to allow for faster simulations with the former when there's no need to
debug. (Could also add it to methods associated with
HistoricalSimulator.__init__(), but those don't produce less output and
are less of a concern.) Also added a method to HistoricalSimulator that
can concurrently plot individual assets' results to visualize how each
performed during the simulation period.