Allowed dividends & empty satellites; squashed bug
Added a new method and associated code to HistoricalSimulator that
looks for dividends in portfolio assets and adds them as cash or partial
shares. (Still need to decide how faithful I want to be to the actual
definition of an ex-date.) Made small additions to both Strategy classes
that allow simulations to proceed when satellite portfolios are empty.
Found an oversight in VolTargetStrategy's calculation of satellite asset
fractions that sometimes over-subtracted and led to negative shares.
I had to change the weight-finding strategy, but it's vulnerable to
floating point errors that make cash go negative maybe .1% of the time.
Using Decimal could fix that, but it's slower than float... I'll think
on it.