Made TopXSplitStrategy; edited core asset handling
Inspired by a sector rotation idea, I wrote a new Strategy that targets
the top performing assets in a past time window. To help it, I edited
the core asset rebalance routine in simulator.py to allow for total
liquidation of core assets (as I already had for satellite assets).