Resolved close price date issue
Completed the work started in the backward-looking commit by making a
class attribute in HistoricalSimulator that gives the previous
iteration's date in begin_time_loop() and using it to retrieve that
close in SMAStrategy instead of peeking ahead at the current day's.

Now, use self.prev_day in similar situations and continue to use a
combination of shift() and self.today with Strategy Series containing
calculated metrics. This change is small enough that it shouldn't change
any test results.